requirements.txt
setuptools-rust
openai-whisper
flask